programmable_dataplanes/pyproject.toml
[project]
name = "programmable-dataplanes"
version = "1.0.0"
requires-python = ">=3.14"
dependencies = [
    "scapy==2.7.0",
    "matplotlib==3.10.9",
    "numpy==2.4.4",
    "scipy==1.17.1",
    "scipy-stubs~=1.17.1",
    "grpcio==1.80.0",
    "grpcio-tools==1.80.0",
    "p4runtime==1.5.0",
]

[dependency-groups]
dev = [
    "ruff",
    "jupyterlab>=4.5.7",
]
